add get and create services to models

+++ app/models/Order.scala
@@ -2,7 +2,7 @@ package models
 
 import play.api.libs.json.Json
 
-case class Order(id: Long, order_date: String, userId: Long)
+case class Order(id: Long, order_date: String, userId: Int)
 
 object Order {
   implicit val orderFormat = Json.format[Order]
